Changed the scope of some of the script-local functions from `incpy#python`.
The `incpy#python` namespace was updated to change the scope for some of
its functionality from script-local to global. This was done so that the
logic inside the `incpy#python` namespace can be split up into a
different namespace where the interpreter and python-specific logic can
be isolated from one another.