Updated the `incpy#python` namespace to consolidate execution logic.
Prior to this commit, the logic for executing internal python code in
the currently available interpreter was split up between the code being
generated and the functions that execute the generated code. This commit
combines both the code generation and its execution into public-scoped
functions that are accessible via the `incpy#python` namespace.

These functions were added alongside the original functionality in order
to allow another commit to transition to these new implementations once
they are actually confirmed working.