Removed the old functions for executing python from the `incpy#python` namespace.
Originally, the logic for executing python in the currently available
interpreter was split up between a function that staged python code
prior to execution, and the function that actually executed it within
the package namespace. That logic has recently been superceded by a
public interface that combines the python code-generation part with the
code that is responsible for executing it.

This is being done prior to separating the interpreter-specific logic
from the internal-interface that is used to interact with an initialized
interpreter.