Added wrappers for setting up keybindings and commands to the `incpy` namespace.
The previous commit copied the utility functions that were in the plugin
script (plugin/incpy.vim) into the `incpy#bindings` autoload namespace.
This commit builds on that by removing said functions from the plugin
script and adding wrappers (with the same names) to the `incpy` autoload
namespace.

A number of other functions were also pulled out of the plugin script in
order to keep all internal functionality inside the "autoload/incpy.vim"
script.