Refactored the current script detection so that it is outside the plugin loader.
On older versions of NeoVim, the 8.2.1297 patch wasn't merged until
later in its lifetime. That patchset resulted in fixing an issue that
would cause expand("<sfile>") to return an incorrect base path when
called by a function that is being executed during initialization.

To avoid the condition that triggers this bug, we grab the current
script path from a different scope when loading the python parts of the
plugin. By grabbing it from a different scope, we work around the
original cause of the issue.